Improved code with the use of protocols
